+++ StreetFighter/.project
@@ -16,7 +16,7 @@
 	</natures>
 	<filteredResources>
 		<filter>
-			<id>1680606451695</id>
+			<id>1680606868102</id>
 			<name></name>
 			<type>30</type>
 			<matcher>
