+++ StreetFighter/.settings/org.eclipse.core.resources.prefs
