+++ StreetFighter/bin/com/skillrisers/streetfighter/sprites/OpponentPlayer.class
