+++ StreetFighter/bin/com/skillrisers/streetfighter/sprites/ken_.png
