+++ StreetFighter/src/com/skillrisers/streetfighter/sprites/OpponentPlayer.java
@@ -1,29 +1,24 @@
 package com.skillrisers.streetfighter.sprites;
 
-import java.awt.Graphics;
 import java.awt.image.BufferedImage;
+
 import javax.imageio.ImageIO;
+
 import com.skillrisers.streetfighter.utils.GameConstants;
 
 public class OpponentPlayer extends CommonPlayer implements GameConstants {
-	private int x;
-	private int y;
-	private int w;
-	private int h;
-	BufferedImage playerImg;
+	
 	public OpponentPlayer() throws Exception {
-		x = SCREENWIDTH - 150;
+		x = SCREENWIDTH - 500;
 		y = GROUND;
 		w = 200;
 		h = 250;
-		playerImg = ImageIO.read(Player.class.getResource("ken_.png"));
+		playerImg = ImageIO.read(Player.class.getResource("ryu_.png"));
 	}
-	public BufferedImage defaultImage(){
-		//
-		return 
-	}
-
-	public void paintPlayer(Graphics pen) {
-		pen.drawImage(playerImg, x, y, w, h, null);
+	
+	@Override
+	public BufferedImage defaultImage() {
+		//X = 2748 Y = 37 Width = 105 Height = 232
+		return playerImg.getSubimage(2748, 37, 105, 232);
 	}
 }
