+++ StreetFighter/src/config.properties
